Rename Epic Client unit test modules/classes based on PEP 0008
